Validation Error

+++ app/Http/Controllers/Country/CountryController.php
@@ -14,7 +14,12 @@ class CountryController extends Controller
     }
 
     public function countryByID($id){
-        return response()->json(CountryModel::find($id),200);
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json("Record not Found!",404);
+        }
+        return response()->json($country,200);
+
     }
 
     public function countrySave(Request $request){
@@ -22,12 +27,20 @@ class CountryController extends Controller
         return response()->json($country,201);
     }
 
-    public function countryUpdate(Request $request, CountryModel $country){
+    public function countryUpdate(Request $request, $id){
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json("Record not Found!",404);
+        }
         $country ->update($request->all());
         return response()->json($country,200);
     }
 
-    public function countryDelete(Request $request, CountryModel $country){
+    public function countryDelete(Request $request, $id){
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json("Record not Found!",404);
+        }
         $country ->delete();
         return response()->json(null,204);
     }
